07.16.Tue ver 0.0.2
몬스터에게 피격 시 데미지 팝업과, 데미지 표기시 단위 축약

+++ Assets/Scripts/PlayerController.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class PlayerController : MonoBehaviour
 {
-    public int att = 10;
+    public long att = 10;
     Animator anim;
     float CurTime = 0;
